backend/pyproject.toml
[tool.poetry]
name = "easy-words"
version = "0.1.0"
description = ""
authors = ["Sergei Sorochinskii <sergei.sorochinskii@dteam.dev>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
sqlalchemy = "^2.0.23"
alembic = "^1.12.1"
uvicorn = "^0.24.0.post1"
pydantic = "^2.5.2"
pydantic-settings = "^2.1.0"
asyncpg = "^0.29.0"
passlib = "^1.7.4"
redis = "^5.0.1"
pyjwt = "^2.8.0"
sqladmin = "^0.16.1"
itsdangerous = "^2.1.2"
authlib = "^1.3.0"
pytest = "^8.1.1"
httpx = "^0.27.0"
trio = "^0.25.0"
anyio = "4.3.0"
fastapi = "^0.110.0"
asgiref = "^3.8.1"
pytest-asyncio = "^0.23.6"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
